// src/types.ts
export interface MediaItem {
  Id: string;
  Title: string;
  CoverImage: string;
  TitleImage: string;
  Date: string;
  ReleaseYear: string;
  MpaRating: string;
  Category: string;
  Duration: string;
  VideoUrl?: string;
  Description: string;
}
export interface DataShape {
  Featured: MediaItem;
  TendingNow: MediaItem[];
}